first commit in a year lets gooo
I had been working on this project for most of this time when I could -- provided real life didn't get in the way. It's certainly been a process, and I've learned a lot though this project. If I may, a word of advice: read ALL the related documentation of what you are using. I lost and had to make up ~6 months of work because I didn't read that Server Sent Events are unidirectional. Had that not happened, this update would have come much sooner.